Bring the new parser in line with the regression tests
In order to incorporate the new parser with the old regression tests, I
needed to update the compiler to C++. When I did this, I also had to
fix a few casting errors that arose due to this change. This is the only
change to the regression tests.

In order to pass (some) of the regression tests, I had to modify the new
parser such that it handles labels correctly. There are still
outstanding issues with the regression tests relating to node order.