Finish the 2 tests that were commented out for newick parser
I finished the 2 tests that I left unfinished from the initial port from
lagrange. The two new tests are

1. manytrees
2. pathologic0

The first tests is just a fuzz test. There is almost no verification on
the results, other than very basic things (like the number of tips is
greater than 0). This is largely implemented to catch any edge cases for
memory leaks, but it still could be better.

The second test is a pathelogic tree that we found at some point. I
think that it broke the old parser, due to it being a very "deep" tree.
This depth caused bison to run out of stack space, and the parse would
fail. I have included in this test to make sure that we don't have the
same error in the future (and to style on bison).