Add `compile_commands.json` to the git ignore.
Clangd, a implementation of the LSP, requires information on how to
compile the project. Either this can be a makefile, or it can be the
compile commands output by cmake. In order to make mine and other's
workflow effective, we need to have this at the project root. So, in
order to keep it from poluting the repository, I have added it to the
gitignore.