Update `corax_random_getint` to use a rejection method
The original method just used mod to compress the range. This method is
generally not very good, as it does not produce a uniform distribution
on the intended range. Most implementations instead use the rejection
method, which just pulls numbers from the generator until a suitable
number is found.

Often, this is slow, so instead a correction is computed so modulus by
the maximum value is uniform. This is what we have done instead. The
algorithm actually implemented is the same as used in the C++ std
library. I pulled most of the information in this commit from [1], which
contains a faster algorithm. However, when implementing it in a toy
environment, I couldn't make sure that the algorithm worked with reduced
wordsize, so I implemented the simpler version instead.

[1]: https://arxiv.org/abs/1805.10941