feat(lbfgsb): disable lb restriction (#123)
The L-BFGS-B code contains a questionable assertion that checks whether the lower bounds are >= CORAX_LBFGSB_ERROR.
This limits the lower bounds to 1e-4 which is too high for optimizing DTL rates in AleRax, for example.
Hence. the assertion was removed and we now dynamically adjust lbfgsb_error based on the lower bound.

Co-authored-by: Noah Wahl <noah.ares@yahoo.de>
Reviewed-on: https://codeberg.org/Exelixis-Lab/coraxlib/pulls/123
Reviewed-by: amkozlov <amkozlov@noreply.codeberg.org>
Co-authored-by: noahares <noahares@noreply.codeberg.org>
Co-committed-by: noahares <noahares@noreply.codeberg.org>